Move UltraMk2 VortexESC frequency configuration to config file
The frequency initialization will be removed. This is going to be
delegated to a caller due to arbitrary groups of pwm being set. If the
pwm pins could have been set individually, this would have been a much
more difference setup; however, adding it in the initialization could
affect other pins which are not used by the motor.