refactor(RobotMecanumLogic): Remove dimensional hack
In order to convert between linear and rotaitonal motion, it must be meters/radians * radius so that the radians cancell out. If you think about the math, this does make dimensional sense. Although radians are dimensionless, it is required to have them in order for units.h to work. Since different dimensionless units can convert into others, this is mathematically desireable at times.