hack(assert): Allow reading most assert from debug mode
If the robot crashes, a breakpoint can be started in main, then the most recent assert be read from memory. Since it is uninitialized, the code will read the most recent info from RAM. In STM32 there is a glitch of ram data being persistent despite being initialized: Noted in turret behavior in RMNA 2025.