refactor(quad_drive): Decouple from taproot subsystems when connected with composition while allowing taproot use
If it only inherits from Subsystem, the initialize and update functionality will only be available as a subsystem. It means that other classes cannot use abstract interfaces which do not care about taproot shenanigans. Although having refresh() and update() doing the same thing is a concern, it gives the option to use a different paradigm from taproot down the line.