test(DualCascadePid): Have proper matching
The test had typos which meant the values were not properly evaluated. For
example, dualPid did not have setTarget called which was a major source for the
deviation from the expected values. Another source of error was the fact that
the max output was zero, which meant that in the initial round of testing, it
passed due to the output from both tests being zero"