fix: Prevent wrapped numbers causing corruption
When utilized for DualCascade GM6020, the wrapped numbers can cause corruption
as they generally cannot be negative. The difference reported should be
positive or negative. When passed to the next PID, it will be able
to scale positively or negatively despite the utilization of wrapped
float algorithms at certain stages.