refactor: Decouple macro shenanigans from header to prevent intellisense errors
Just to develop with some grammar checking, intellisense needs a specific robot
chosen. Instead of having to favor a specific robot build which may lead to
confusion down the line, the macros are moved into the translation unit while
the header file is "normal". The Singleton pattern also makes the construction
of the robot that is used within main() a bit more cleaner.

Note that Intellisense does not attempt to anticipate linker errors.