fix: Syntax namespace mangling
Whenever the includes were done within a namespace, it mangled EVERY include
that followed afterwards. So any namespace std:: would be within fang::std
But not eveyr std:: is included that those std's rely on at times which leads
to the compiler creating a half-made duplicate set of std.

Weighing out the consequences of macros to support seven robots includes
the measures that are taken to isolate such shenanigans. In this case, the
shenanigans were removed from the header file, which declared some global variable,
and variable macro types.