fix: Test segmentation fault triggered by taproot
Taproot has a bug in DjiMotorEncoderMock. This is now triggered because the
actual robots are being instantiated during the unit test builds due to the
removal of their definitions from main(). Since they are just part of your
average translation unit, it happens to be built.

A future solution could be adding it to the exlusion file for unit tests.

For now, this will do.