fix: Config variables being zero on hardware builds
This was due to a static variable fiasco since we don't have
C++ modules. The 2nd construct on first use idiom has been
chosen due to the fact that the robot config should exist through the
whole program like a good static variable.

In the future this allows for some weird shenanigans to allow
these large static variables to not be copied but destroyed once
they are done being used.

We could also use an rvalue constructor but this is unnecessary at the moment