fix: Flywheels not being activated
It turns out that makeVectorWithMovedList was not triggering the lambda within
the fold expression.  This was because the lambda statement was not being
called. For those unfamiiliar treat the portion of the lambda statement itself
as the "function name". In order for it to be called you need to have the ()
with the argumenst afterwards which does lookk a bit odd.

https://iifx.dev/en/articles/457564524/a-friendly-guide-to-fold-expressions-and-lambdas-in-c-17